erp/requirements.txt
Flask==3.1.2
python-dateutil==2.9.0.post0